Update VYNFCNDEFPayloadParser.m
This change handles issue https://github.com/vinceyuan/VYNFCKit/issues/4 - to ignore unknown attribute tags in wifi simple configuration parser.

There is a low change of conflict in the code because unknown attribute can include additional attribute information of variable length. If this includes a byte sequence wich is equal to an attribute tag, the parser would misinterpret this as an attribute!